Fix warning C4267

+++ TeamProjects/diffusion-solver/core/src/SchemeLayerGeneratorInitial.cpp
@@ -3,7 +3,7 @@
 using namespace diffusioncore;
 
 const double PI_CONST = 3.14159265358979323846;
-inline double EvaluateCos(double x, int index) {
+inline double EvaluateCos(double x, size_t index) {
    return std::cos(PI_CONST * x * index);
 }
